Celestial Static Method Additions
* Add ability to grab last / next celestial rises and sets from specified points in time.
