+++ CoordinateSharp/CoordinateSharp.csproj
@@ -54,7 +54,9 @@ For more information, please contact Signature Group, LLC at this address: sales
     <PackageLicenseUrl></PackageLicenseUrl>
     <Copyright>Copyright 2020</Copyright>
     <Description>A simple library designed to assist with geographic coordinate conversions, formatting and location based solar and lunar information.</Description>
-    <PackageReleaseNotes>Adds a standard parse method that will throw exceptions on failure. Various XML improvements.</PackageReleaseNotes>
+    <PackageReleaseNotes>* Adds a standard parse method that will throw exceptions upon failure. 
+* Adds ability to get last or next celestial body rise or set via static methods. 
+* Various XML improvements.</PackageReleaseNotes>
     <PackageTags>CoordinateSharp Latitude Longitude Coordinates Geography Sun Moon Solar Lunar Time MGRS UTM Julian ECEF Sunset Sunrise Set Moonset Moonrise</PackageTags>
     <!-- <PackageLicenseExpression>AGPL-3.0-or-later</PackageLicenseExpression>-->
     <PackageLicenseFile>License.txt</PackageLicenseFile>    <PackageIconUrl></PackageIconUrl>
